Final updates to crud stuff for arquillian test configs
