+++ crud/src/test/java/com/redhat/lightblue/rest/crud/hystrix/AbstractRestCommandTest.java
@@ -18,7 +18,10 @@
  */
 package com.redhat.lightblue.rest.crud.hystrix;
 
+import com.netflix.hystrix.util.ExceptionThreadingUtility;
 import com.redhat.lightblue.Response;
+import com.redhat.lightblue.config.CrudConfiguration;
+import com.redhat.lightblue.config.MetadataConfiguration;
 import com.redhat.lightblue.crud.DeleteRequest;
 import com.redhat.lightblue.crud.Factory;
 import com.redhat.lightblue.crud.FindRequest;
@@ -27,13 +30,49 @@ import com.redhat.lightblue.crud.SaveRequest;
 import com.redhat.lightblue.crud.UpdateRequest;
 import com.redhat.lightblue.mediator.Mediator;
 import com.redhat.lightblue.metadata.Metadata;
+import com.redhat.lightblue.rest.RestConfiguration;
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.EmptyAsset;
+import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.jboss.shrinkwrap.resolver.api.maven.Maven;
 import org.junit.After;
+import org.junit.runner.RunWith;
+
+import java.io.File;
 
 /**
  *
  * @author nmalik
  */
+@RunWith(Arquillian.class)
 public abstract class AbstractRestCommandTest {
+    private static final String CONFIGPROPERTIES = "config.properties";
+
+    @Deployment
+    public static WebArchive createDeployment() {
+        ExceptionThreadingUtility.assignCallingThread(Thread.currentThread());
+
+        File[] libs = Maven.resolver().loadPomFromFile("pom.xml").importRuntimeDependencies().resolve().withTransitivity().asFile();
+
+        final String PATH = "src/test/resources/it/it-";
+
+        WebArchive archive = ShrinkWrap.create(WebArchive.class, "test.war")
+                .addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml")
+                .addAsResource(new File(PATH + MetadataConfiguration.FILENAME), MetadataConfiguration.FILENAME)
+                .addAsResource(new File(PATH + CrudConfiguration.FILENAME), CrudConfiguration.FILENAME)
+                .addAsResource(new File(PATH + RestConfiguration.DATASOURCE_FILENAME), RestConfiguration.DATASOURCE_FILENAME)
+                .addAsResource(new File(PATH + CONFIGPROPERTIES), CONFIGPROPERTIES);
+
+        for (File file : libs) {
+            archive.addAsLibrary(file);
+        }
+        archive.addPackages(true, "com.redhat.lightblue");
+        return archive;
+
+    }
+
     public static class TestMediator extends Mediator {
         public String methodCalled;
         public Object[] args;
