+++ crud/src/test/resources/ITCaseCrudResourceTest/config/config.properties
@@ -1,4 +1,3 @@
 # set default isolation strategy for tests to SEMAPHORE else breakpoints in IDE
 # when debugging unit tests will timeout with default strategy (THREAD).
 hystrix.command.default.execution.isolation.strategy=SEMAPHORE
-
