+++ crud/src/test/resources/ITCaseCrudResourceTest/config/datasources.json
@@ -1,0 +1,14 @@
+{
+   "mongo" : {
+     "metadataBackendParser" : "com.redhat.lightblue.metadata.mongo.MongoBackendParser",
+     "ssl" : false,
+     "database" : "mongo",
+     "servers" : [
+        {
+           "host" : "localhost",
+           "port" : "27757"
+        }
+     ],
+     "type" : "com.redhat.lightblue.mongo.config.MongoConfiguration"
+   }
+}
