+++ crud/src/test/resources/ITCaseCrudResourceTest/config/lightblue-metadata.json
@@ -1,0 +1,5 @@
+{
+    "type" : "com.redhat.lightblue.mongo.config.MongoMetadataConfiguration",
+    "dataSource" : "mongo",
+    "collection": "metadata"
+}
