+++ crud/src/test/resources/ITCaseCrudResourceTest/expectedDeleted.json
