+++ crud/src/test/resources/ITCaseCrudResourceTest/expectedFound.json
