+++ crud/src/test/resources/ITCaseCrudResourceTest/expectedInserted.json
@@ -1,0 +1,13 @@
+{
+    "status": "COMPLETE",
+    "modifiedCount": 1,
+    "matchCount": 0,
+    "processed": [
+        {
+            "iso3code": "CAN",
+            "name": "Canad",
+            "iso2code": "CA",
+            "objectType": "country"
+        }
+    ]
+}
