+++ crud/src/test/resources/ITCaseCrudResourceTest/resultDeleted.json
