+++ crud/src/test/resources/ITCaseCrudResourceTest/resultUpdated.json
